Refactor and move files to folders for clarity
