+++ front-end/src/FotoSoppi.jsx
@@ -1,13 +1,13 @@
 import * as React from 'react';
-import FotoSoppi from './pages/FotoSoppi';
+import Home from './pages/Home';
 import Login from './pages/Login';
 import { Routes, Route} from 'react-router-dom'
 
 
-export default function App() {
+export default function FotoSoppi() {
   return (
     <Routes>
-      <Route path="/" element={<FotoSoppi />} />
+      <Route path="/" element={<Home />} />
       <Route path="/login" element={<Login />} />
     </Routes>
   );
