+++ front-end/src/pages/components/ImageCardComponent.jsx
